Adding ability to pan to highest risk area
A new route has been added to the backend server that will
query the database looking for the highest population polygon
that is at risk for zika and does not have care delivery.  The
map will then pan to the location of that polygon.

Updated div layouts to accomodate the button and make things
more responsive.