Adding layer toggle button and refactoring map.
Adding action and store for layer state.  It can either be
visible or invisible.  The button is currently in the
RiskMenu component.  Maybe we should move this later?

I've also refactored the Map component.  The polygon layer has
been removed and placed in it's own component `VectorLayer`.
This makes it much easier to mount an unmount only this layer.